+++ test/python/symmetrize_gf.py
@@ -35,7 +35,7 @@ wmesh = MeshImFreq(beta=p.beta, S='Fermion', n_max=p.nw)
 cell = np.eye(3)
 bl = BravaisLattice(cell)
 bz = BrillouinZone(bl)
-kmesh = MeshBrZone(bz, p.nk * np.eye(3, dtype=np.int32))
+kmesh = MeshBrZone(bz, p.nk * np.eye(3, dtype=int))
 
 gf = Gf(mesh=MeshProduct(wmesh, kmesh), target_shape=2*(p.norb,))
 gf.data[:] = np.random.rand(*gf.data.shape)
